pyproject.toml
[tool.poetry]
name = "ga_to_1pw"
version = "0.1.0"
description = ""
authors = ["aloysius <alloystory@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.3.0"
pyzbar = "^0.1.9"
protobuf = "^4.21.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
